Remove enum in PnlVectCompact
The definition of the PnlVectCompact type is modified a little to ensure
that the array field is set to NULL whenever the content is stored in a
compact format. To make it possible, we had to ensure that both the val and
array fields exist on their own and not inside an enum.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/pnl/trunk@968 94367915-3dfb-0310-9a54-a1fefb3e5df9