Add or Remove elements from bases
Onece a basis is created, all its elements are described by the tensor
matrix. These two functions enable to add or delete a line in the tensor so
that the user can remove an element from the basis or add a new one bu
described in terms of powers of the initial 1d basis.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/pnl/trunk@974 94367915-3dfb-0310-9a54-a1fefb3e5df9