Bases use the sparse tensor to speed up evaluation
We add an extra memeber SpT in the PnlBasis, which is a sparse
representation of the tensor matrix T. The introduction of this sparse
storage enables to speed up the evaluation of multi-dimensional bases by a
great deal; the computational time is divided by (nb_variates/degree).

For the moment, the function pnl_basis_eval_derivs does not use this sparse
representation.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/pnl/trunk@1019 94367915-3dfb-0310-9a54-a1fefb3e5df9