Improve the creation of basis with a sum degree
When the total total is given as the sum of the partial degrees, we know
the total number of terms involved in the multivariate polynomials : (n + d
choose d) for a n-variate polynomial with total degree d. So, we can
directly build the tensor with the proper size without making copies each
time we add a variate.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/pnl/trunk@1100 94367915-3dfb-0310-9a54-a1fefb3e5df9